+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* *Upcoming Release*
+  * Fix dynamic properties not working with archive installations. [#110](https://github.com/update4j/update4j/issues/110)
 * **1.5.4**
   * Fix archive not properly resolving UNIX paths. [#109](https://github.com/update4j/update4j/issues/109)
 * **1.5.3**